+++ R/methods.R
@@ -26,7 +26,7 @@ logLik.logitr <- function(object, ...) {
 #' @rdname miscmethods.logitr
 #' @export
 terms.logitr <- function(x, ...) {
-  return(x$inputs$pars)
+  return(stats::terms(x$formula))
 }
 
 #' @rdname miscmethods.logitr
@@ -447,4 +447,106 @@ residuals.logitr <- function(object, fitted = NULL, ...) {
   residuals$fitted_value <- NULL
   residuals$residual <- as.vector(resids)
   return(residuals)
+}
+
+#' Extract Model Confidence Interval
+#'
+#' Returns confidence intervals from an object of class `logitr`.
+#' @keywords logitr confint
+#'
+#' @param object is an object of class `logitr` (a model estimated using
+#' the 'logitr()` function).
+#' @param parm A specification of which parameters are to be given confidence
+#' intervals, either a vector of numbers or a vector of names.
+#' If missing, all parameters are considered.
+#' @param level The confidence level required.
+#' @param ... further arguments.
+#'
+#' @return A data frame of the confidence intervals of model coefficients.
+#' @export
+#' @examples
+#' library(logitr)
+#'
+#' # Estimate a preference space model
+#' mnl_pref <- logitr(
+#'   data    = yogurt,
+#'   outcome = "choice",
+#'   obsID   = "obsID",
+#'   pars    = c("price", "feat", "brand")
+#' )
+#'
+#' # Compute a confidence interval
+#' confint(mnl_pref)
+confint.logitr <- function(object, parm, level = 0.95, ...) {
+    draws <- getUncertaintyDraws(object, numDraws = 10^4)
+    lower <- (1 - level)/2
+    upper <- 1 - lower
+    df <- data.frame(
+        lower = apply(draws, 2, function(x) fquantile(x, lower, na.rm = TRUE)),
+        upper = apply(draws, 2, function(x) fquantile(x, upper, na.rm = TRUE))
+    )
+    names(df) <- c(paste(lower*100, "%"), paste(upper*100, "%"))
+    return(df)
+}
+
+#' Construct Design Matrices
+#'
+#' Creates a design (or model) matrix, e.g., by expanding factors to a set of
+#' dummy variables (depending on the contrasts) and expanding interactions
+#' similarly.
+#' @keywords logitr model.matrix
+#'
+#' @param object an object of an appropriate class. For the default method,
+#' a model `formula` or a `terms` object.
+#' @param ... further arguments.
+#'
+#' @return A design matrix
+#' @export
+#' @examples
+#' library(logitr)
+#'
+#' # Estimate a preference space model
+#' mnl_pref <- logitr(
+#'   data    = yogurt,
+#'   outcome = "choice",
+#'   obsID   = "obsID",
+#'   pars    = c("price", "feat", "brand")
+#' )
+#'
+#' # Get the model.matrix design matrix
+#' model.matrix(mnl_pref)
+model.matrix.logitr <- function(object, ...) {
+    return(object$data$X)
+}
+
+#' Extracting the Model Frame from a Formula or Fit
+#'
+#' Returns a data.frame with the variables needed to use formula and
+#' any `...` arguments.
+#' @keywords logitr model.frame
+#'
+#' @param formula a model `formula` or `terms` object or an R object.
+#' @param ... further arguments.
+#'
+#' @return A data.frame with the variables needed to use formula and
+#' any `...` arguments.
+#' @export
+#' @examples
+#' library(logitr)
+#'
+#' # Estimate a preference space model
+#' mnl_pref <- logitr(
+#'   data    = yogurt,
+#'   outcome = "choice",
+#'   obsID   = "obsID",
+#'   pars    = c("price", "feat", "brand")
+#' )
+#'
+#' # Get the model.frame data frame
+#' model.frame(mnl_pref)
+model.frame.logitr <- function(formula, ...) {
+    stats::model.frame.default(
+        formula$formula,
+        data = eval(formula$call$data, envir = parent.frame())
+    )
 }