+++ tests/testthat/test_startVals.R
@@ -1,7 +1,7 @@
 context("Starting parameters")
 library(logitr)
 
-test_that("getStartPars returns all 0s for first iteration in a multistart loop (and not other iterations)", {
+test_that("getStartVals returns all 0s for first iteration in a multistart loop (and not other iterations)", {
   modelInputs <- getModelInputs(
     data = yogurt,
     outcome = "choice",
@@ -14,7 +14,7 @@ test_that("getStartPars returns all 0s for first iteration in a multistart loop 
     panelID = NULL,
     clusterID = NULL,
     robust = FALSE,
-    startParBounds = c(-1, 1),
+    startValBounds = c(-1, 1),
     startVals = NULL,
     numMultiStarts  = 1,
     useAnalyticGrad = TRUE,
@@ -37,15 +37,15 @@ test_that("getStartPars returns all 0s for first iteration in a multistart loop 
       algorithm   = "NLOPT_LD_LBFGS"
     )
   )
-  startPars1 <- getStartPars(modelInputs, i = 1)
-  startPars2 <- getStartPars(modelInputs, i = 2)
-  expect_identical(startPars1, c(price = 0, feat = 0))
-  expect_false(identical(startPars2, c(price = 0, feat = 0)))
-  expect_equal(length(startPars1), 2)
-  expect_equal(length(startPars2), 2)
+  startVals1 <- getStartVals(modelInputs, i = 1)
+  startVals2 <- getStartVals(modelInputs, i = 2)
+  expect_identical(startVals1, c(price = 0, feat = 0))
+  expect_false(identical(startVals2, c(price = 0, feat = 0)))
+  expect_equal(length(startVals1), 2)
+  expect_equal(length(startVals2), 2)
 })
 
-test_that("getStartPars returns user-provided starting parameters for first iteration in a multistart loop (and not other iterations)", {
+test_that("getStartVals returns user-provided starting parameters for first iteration in a multistart loop (and not other iterations)", {
   modelInputs <- getModelInputs(
     data = yogurt,
     outcome = "choice",
@@ -58,7 +58,7 @@ test_that("getStartPars returns user-provided starting parameters for first iter
     panelID = NULL,
     clusterID = NULL,
     robust = FALSE,
-    startParBounds = c(-1, 1),
+    startValBounds = c(-1, 1),
     startVals = c(1, 1),
     numMultiStarts = 1,
     useAnalyticGrad = TRUE,
@@ -81,10 +81,10 @@ test_that("getStartPars returns user-provided starting parameters for first iter
       algorithm   = "NLOPT_LD_LBFGS"
     )
   )
-  startPars1 <- getStartPars(modelInputs, i = 1)
-  startPars2 <- getStartPars(modelInputs, i = 2)
-  expect_identical(startPars1, c(price = 1, feat = 1))
-  expect_false(identical(startPars2, c(price = 0, feat = 0)))
-  expect_equal(length(startPars1), 2)
-  expect_equal(length(startPars2), 2)
+  startVals1 <- getStartVals(modelInputs, i = 1)
+  startVals2 <- getStartVals(modelInputs, i = 2)
+  expect_identical(startVals1, c(price = 1, feat = 1))
+  expect_false(identical(startVals2, c(price = 0, feat = 0)))
+  expect_equal(length(startVals1), 2)
+  expect_equal(length(startVals2), 2)
 })